# Replace readItem in ProcessNumber to speed-up SolarShadingTest_ImportedShading