--- a/src/EnergyPlus/UtilityRoutines.cc
+++ b/src/EnergyPlus/UtilityRoutines.cc
@@ -135,13 +135,12 @@ namespace UtilityRoutines {
             // then parse as a normal floating point value
             try {
                 rProcessNumber = std::stod(PString, nullptr);
-                ErrorFlag = false;
-            } catch (std::invalid_argument &e) {
-                ErrorFlag = true;
-                rProcessNumber = 0.0;
+            } catch (std::invalid_argument &e) { // LCOV_EXCL_LINE
+                rProcessNumber = 0.0;            // LCOV_EXCL_LINE
+                ErrorFlag = true;                // LCOV_EXCL_LINE
             } catch (std::out_of_range &e) {
-                ErrorFlag = true;
                 rProcessNumber = 0.0;
+                ErrorFlag = true;
             }
         } else {
             rProcessNumber = 0.0;