--- a/tst/EnergyPlus/unit/UtilityRoutines.unit.cc
+++ b/tst/EnergyPlus/unit/UtilityRoutines.unit.cc
@@ -216,43 +216,49 @@ TEST_F(EnergyPlusFixture, UtilityRoutines_ProcessNumber)
     double expectedVal{3.14159};
     bool expectedError{false};
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(goodString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, false);
+    EXPECT_FALSE(expectedError);
 
     goodString = "3.14159+E0";
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(goodString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, false);
+    EXPECT_FALSE(expectedError);
 
     goodString = "3.14159+e0";
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(goodString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, false);
+    EXPECT_FALSE(expectedError);
 
     goodString = "3.14159+D0";
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(goodString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, false);
+    EXPECT_FALSE(expectedError);
 
     goodString = "3.14159+d0";
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(goodString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, false);
+    EXPECT_FALSE(expectedError);
 
     // invalid strings
     std::string badString{"É.14159"};
     expectedVal = 0.0;
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(badString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, true);
+    EXPECT_TRUE(expectedError);
 
     badString = "3.14159É0";
     expectedVal = 0.0;
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(badString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, true);
+    EXPECT_TRUE(expectedError);
 
     badString = "3.14159 0";
     expectedVal = 0.0;
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(badString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, true);
+    EXPECT_TRUE(expectedError);
+
+    // invalid argument
+    badString = "E3.14159";
+    expectedVal = 0.0;
+    EXPECT_NEAR(UtilityRoutines::ProcessNumber(badString, expectedError), expectedVal, 1E-5);
+    EXPECT_TRUE(expectedError);
 
     // out of range
     badString = "1E5000";
     expectedVal = 0.0;
     EXPECT_NEAR(UtilityRoutines::ProcessNumber(badString, expectedError), expectedVal, 1E-5);
-    EXPECT_EQ(expectedError, true);
+    EXPECT_TRUE(expectedError);
 }