@jmythms 2021-06-18
The run time differences are amazing. I wish we knew what was causing so much improvement in Mac builds, or, what was causing Mac builds to slow down so much with the original code. The runtimes in Linux builds only had an improvement from 8s to 5s. I wonder if it is something to do with the compiler we use for Mac builds, since the code is the same.