@mitchute 2021-06-18
I'm not sure what the issue is. ```readItem``` [is a templated function](https://github.com/NREL/EnergyPlus/blob/5358d3eb87d1376b3484e1a38ade8274b01f0027/src/EnergyPlus/StringUtilities.hh#L79-L84), so maybe that has something to do with it. Maybe in the future we can move away from ```readItem``` for parsing strings. What did you do in the ScheduleManager refactor?