@jmythms 2021-06-18
The ScheduleManager refactor is focused on decreasing unnecessary [file open](https://github.com/NREL/EnergyPlus/blob/5358d3eb87d1376b3484e1a38ade8274b01f0027/src/EnergyPlus/ScheduleManager.cc#L1792), close operations where they can be avoided. I was using the original ProcessNumber in it. The improvements from this PR will definitely add to the effects of the ScheduleManager refactor.