@jmythms 2021-06-18
Since ProcessNumber is used in three places in [ScheduleManager.cc](https://github.com/NREL/EnergyPlus/blob/develop/src/EnergyPlus/ScheduleManager.cc), any case that uses those parts of the code should see improvements from this PR:
1. Schedule:File:Shading
2. Schedule:Compact
3. Schedule:File