@mitchute 2021-06-18
Ah, I figured out the ```invalid_argument``` exception. "E3.14159" will fall in there. I'll fix that up.

Anything else?