@rraustad 2021-06-18
> OK, the unit test has been added. I marked the `std::invalid_argument` exception to be skipped for coverage results. I can't make it fall in there with any of the following:
> 
> "3.14159+0" - returns 3.14159
> "3.14159.0" - returns 3.14159
> "3.14159+-0" - returns 3.14159
> "3.14159-E-0" - returns 3.14159
> 
> Maybe we should also be checking for multiple "+", "-", ".", and "E" (or similar, "d", "D", or "e"). It looks like `std::stod` just reads up to the first place where it finds a valid number and returns that.

With this speed improvement, wouldn't it be smarter to catch that in the input processor?