@mitchute 2021-06-18
> wouldn't it be smarter to catch that in the input processor?

I'm honestly not sure what the right path is currently. I think we'll find that this a part of *all* input processing, regardless of whether it happens in the ```InputProcessor```. ```readItem``` is used explicitly all over the place to read in files, including the weather files. I was just looking at a file where my local profiler was saying that ```readItem``` was taking something like 20% of a 15-second simulation. We ought to take a look at seeing where we can eliminate it in other places.