@rraustad 2021-06-18
Just a thought, could this be moved to the input processor for numeric inputs? That certainly does not have to happen in the branch since all looks good here.

            std::replace_if(
                std::begin(PString), std::end(PString), [](const char c) { return c == 'D' || c == 'd'; }, 'e');