@jmythms 2021-06-21
This looks good to me. Testing done:

1. Checked with Fortran "numerics" like "2d3", the parsing worked well: Converted 2d3 to 2e3 and then to 2000.  
2. Strings like "asdf" changed the ErrorFlag to true.